% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_diagnostics.R
\name{accuracy_G_P_diagnostic}
\alias{accuracy_G_P_diagnostic}
\title{accuracy_G_P_diagnostic}
\usage{
accuracy_G_P_diagnostic(P_a, G_a, Initial_Data)
}
\arguments{
\item{P_a}{?}

\item{G_a}{?}

\item{Initial_Data}{simulated epidemic data}
}
\value{
print(mean(false_positives_dist_P)), print(mean(true_positives_dist_P)), print(mean(false_positives_dist_G)), print(mean(true_positives_dist_G)), mean of true and false positives for contact and transmission networks
}
\description{
give statistics on contact and transmission network
}
