% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_sim.R
\name{Bayes_Net_Inf_CCM}
\alias{Bayes_Net_Inf_CCM}
\title{Bayes_Net_Inf_CCM}
\usage{
Bayes_Net_Inf_CCM(
  population = 1000,
  beta_a = 0,
  beta_l = (1 - (1 - 0.128)^(1/14)),
  gamma_a = 1/4,
  gamma_l = 1/14,
  lambda = 10,
  degree_dist = "poisson",
  num_samples = 25,
  genetic_bits = 1024,
  n_mcmc_trials = 2000,
  MCMC_wgt = 0.75,
  ratio_orig_mean_deg = 1,
  epidemic = NULL
)
}
\arguments{
\item{population}{number of individuals in the network}

\item{beta_a}{probability of infection for acute phase}

\item{beta_l}{probability of infection for long-term phase}

\item{gamma_a}{length of infection for acute phase}

\item{gamma_l}{length of infection for long-term phase}

\item{lambda}{mean degree for the Poisson degree distribution}

\item{degree_dist}{degree distribution of contacts, can be "poisson" (requires lambda parameter) or "nbinom" (requires size and mu parameters)}

\item{num_samples}{number of behavior survey samples}

\item{genetic_bits}{length of genetic sequence}

\item{n_mcmc_trials}{number of MCMC trials to perform}

\item{MCMC_wgt}{weight from 0 to 1 indicating how much weight to put on the prior versus the MCMC weight}

\item{ratio_orig_mean_deg}{ratio of how much to inflate or deflate number of partners reported; is no increase/decrease, 0.5 is half the mean degree reported, 2 is twice the mean degree reported}

\item{epidemic}{select a set of epidemic parameters to use instead of entering custom parameters, can be set to "HIV" or "COVID" (overrides and parameters set in the function's arguments); set to "NULL" to use custom parameters via the function's arguments}

\item{size}{scale for the Negative Binomial degree distribution}

\item{mu}{mean degree for the Negative Binomial degree distribution}
}
\value{
results_list = list(G_stats.df, Prob_Distr_Params, Prob_Distr_Params_hyperprior, n_mcmc_trials, G_stats_truth, init_seed), results of data simulation and Bayesian inference, including estimated contact network statistics
}
\description{
generates data and estimates contact network properties using survey, clinical, and genetic data, reads parameters froo read_params()

sets parameters for simulating network data and performing Bayesian inference, which is run with Bayes_Net_Inf_CCM()
}
