% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesnetinf-package.R
\docType{package}
\name{bayesnetinf-package}
\alias{bayesnetinf}
\alias{bayesnetinf-package}
\title{bayesnetinf: Estimate Network Parameters with CCMs and Bayesian Inference}
\description{
A framework to estimate parameters of a static, agent-based network model. Parameters of the contact network are estimated with the use of Bayesian inference and Congruence Class Models (CCMs) in order to incorporate multiple data sources (epidiological, viral genetic, and risk behavior survey data).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ravigoyalgit/Bayes_Net_Inf_CCM}
  \item Report bugs at \url{https://github.com/ravigoyalgit/Bayes_Net_Inf_CCM/issues}
}

}
\author{
\strong{Maintainer}: Sally K. Slipher \email{sallyslipher@montana.edu}

Authors:
\itemize{
  \item Ravi Goyal \email{r1goyal@health.ucsd.edu}
}

}
\keyword{internal}
