% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_gen_epi_data.R
\name{gen_reporting_bias}
\alias{gen_reporting_bias}
\title{gen_reporting_bias}
\usage{
gen_reporting_bias(sampled_v_deg)
}
\arguments{
\item{sampled_v_deg}{reported partners before simulating bias}
}
\value{
bias_sampled_v_deg, the simulated, biased reported number of partners
}
\description{
adjust reported partners by the ratio_orig_mean_deg to simulate bias
}
