% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_gen_epi_data.R
\name{generate_epidemic_data}
\alias{generate_epidemic_data}
\title{generate_epidemic_data}
\usage{
generate_epidemic_data(
  beta_a = 1/5,
  beta_l = 1/2.5,
  gamma_a = 1/10,
  gamma_l = 1/100,
  population = 500,
  Prob_Distr_Params = list(list(rep(1, 10))),
  Prob_Distr = "DirMult",
  num_init_infected = 1,
  genetic_bits = 1024,
  Network_stats = NULL,
  strong_prior = NULL,
  num_samples = 100,
  covPattern = NULL
)
}
\arguments{
\item{beta_a}{= probability of infection for acute phase (default: 1/5)}

\item{beta_l}{= probability of infection for long-term phase (default: 1/2.5)}

\item{gamma_a}{= length of infection for acute phase (default: 1/10)}

\item{gamma_l}{= length of infection for long-term phase (default: 1/100)}

\item{population}{= number of individuals in the network (default: 500)}

\item{Prob_Distr_Params}{= (default: list(list(rep(1,10))))}

\item{Prob_Distr}{= (default: 'DirMult')}

\item{num_init_infected}{= number of individuals infected at the start of the epidemic (default: 1)}

\item{genetic_bits}{= length of genetic sequence (default: 1024)}

\item{Network_stats}{= (default: NULL)}

\item{strong_prior}{= (default: NULL)}

\item{num_samples}{= number of behavior survey samples (default: 100)}

\item{covPattern}{= (default: NULL)}
}
\value{
"No Epidemic: restarting..."; list(G, P, Ia, Il, R, T_dist, Prob_Distr_Params_hyperprior, Prob_Distr_Params, CCMnet_Result[\link{2}]) or "ERROR: FAILED", simulated data, including contact and transmission networks, transition times, and genetic distances
}
\description{
simulates epidemic data from epidemic parameters
}
