% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_gen_genetic_data.R
\name{Genetic_Seq_Data}
\alias{Genetic_Seq_Data}
\title{Genetic_Seq_Data}
\usage{
Genetic_Seq_Data(
  P,
  Ia,
  final_time = max(Ia[which(Ia < Inf)]),
  genetic_bits = 128
)
}
\arguments{
\item{P}{transmission network}

\item{Ia}{vector of times of infection for each individual}

\item{final_time}{latest observed infection time}

\item{genetic_bits}{length of genetic sequences to generate}
}
\value{
Genetic_Seq_mat, matrix of genetic sequences; rows are individuals, columns are genetic bits
}
\description{
generates phylogenetic data by assigning genetic sequences to all individuals who are infected at the end of simulation
}
