% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_diagnostics.R
\name{genetic_seq_diagnostic}
\alias{genetic_seq_diagnostic}
\title{genetic_seq_diagnostic}
\usage{
genetic_seq_diagnostic(P, T)
}
\arguments{
\item{P}{transmission network}

\item{T}{viral genetic network}
}
\value{
list(mean(dist_linked_cases), mean(T_data\link{which(T_data > 0)}), var(dist_linked_cases), var(T_data\link{which(T_data > 0)})), mean and variance of genetic sequence distances
}
\description{
gives statistics on genetic sequence distances
}
