% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_gen_genetic_data.R
\name{genetic_seq_parent_node}
\alias{genetic_seq_parent_node}
\title{genetic_seq_parent_node}
\usage{
genetic_seq_parent_node(Genetic_Seq_a, last_updatetime, ctime, genetic_bits)
}
\arguments{
\item{Genetic_Seq_a}{genetic sequence that is inherited, to be mutated}

\item{last_updatetime}{last update time}

\item{ctime}{current time}

\item{genetic_bits}{length of genetic sequences to generate}
}
\value{
Genetic_Seq_a, a mutated genetic sequence
}
\description{
generates a genetic sequence for an infected individual by mutating genetic bits proportional to how long infected
}
