% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_gen_epi_data.R
\name{Initialize_G_P_Ia_Il_R}
\alias{Initialize_G_P_Ia_Il_R}
\title{Initialize_G_P_Ia_Il_R}
\usage{
Initialize_G_P_Ia_Il_R(
  population,
  beta_a,
  beta_l,
  gamma_a,
  gamma_l,
  G = NULL,
  num_init_infected = 1
)
}
\arguments{
\item{population}{= number of individuals in the network}

\item{beta_a}{probability of infection for acute phase}

\item{beta_l}{probability of infection for long-term phase}

\item{gamma_a}{length of infection for acute phase}

\item{gamma_l}{length of infection for long-term phase}

\item{G}{contact network (default: NULL)}

\item{num_init_infected}{= number of individuals infected at the start of the epidemic (default: 1)}
}
\value{
list(G,P,Ia,Il,R), the initialized contact and transmission networks and transition times
}
\description{
initializes the contact and transmission networks and transition times
}
