% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_diagnostics.R
\name{network_properties_diagnostic}
\alias{network_properties_diagnostic}
\title{network_properties_diagnostic}
\usage{
network_properties_diagnostic(G_a, burnin, view_graphs = FALSE)
}
\arguments{
\item{G_a}{?}

\item{burnin}{length of burn-in}

\item{view_graphs}{plot network property assortativity and density (default: FALSE)}
}
\value{
list(length(which(assort_a >=  assort_truth))/length(assort_a), mean(assort_a), assort_truth, mean(ecount_a), ecount_truth), means, true values, and proportions for edge count and assortativity
}
\description{
gives statistics on network properties
}
