% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_gen_epi_data.R
\name{SIIR.simulator}
\alias{SIIR.simulator}
\title{SIIR.simulator}
\usage{
SIIR.simulator(
  g,
  population,
  beta_a,
  beta_l,
  gamma_a,
  gamma_l,
  num_init_infected = 1
)
}
\arguments{
\item{g}{?}

\item{population}{= number of individuals in the network}

\item{beta_a}{probability of infection for acute phase}

\item{beta_l}{probability of infection for long-term phase}

\item{gamma_a}{length of infection for acute phase}

\item{gamma_l}{length of infection for long-term phase}

\item{num_init_infected}{= number of individuals infected at the start of the epidemic (default: 1)}
}
\value{
exampleepidemic, with which nodes are infected, their parent, and their transition times (acute, long-term, and treatment)
}
\description{
simulates a SIIR epidemic process on a population
}
