% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_diagnostics.R
\name{uncategorized_diagnostic}
\alias{uncategorized_diagnostic}
\title{uncategorized_diagnostic}
\usage{
uncategorized_diagnostic(P_a, G_a, G, P, Ia, Il, R, Initial_Data)
}
\arguments{
\item{P_a}{?}

\item{G_a}{?}

\item{G}{contact network}

\item{P}{transmission network}

\item{Ia}{vector of times of infection for each individual}

\item{Il}{vector of times of transition to long-term infection for each individual}

\item{R}{vector of times of transition to recovery for each individual}

\item{Initial_Data}{simulated epidemic data}
}
\value{
print(cor(Ia, Initial_Data[\link{3}])), print(cor(Il, Initial_Data[\link{4}])), correlations of infection data (acute and long-term) with those simulated
}
\description{
gives statistics on infection times (acute and long-term)
}
