% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_bayes_inf.R
\name{Update_G}
\alias{Update_G}
\title{Update_G}
\usage{
Update_G(
  G,
  P,
  Ia,
  Il,
  R,
  beta_a,
  beta_l,
  Prob_Distr_Params,
  Network_stats,
  Prob_Distr,
  covPattern
)
}
\arguments{
\item{G}{contact network}

\item{P}{transmission network}

\item{Ia}{vector of times of infection for each individual}

\item{Il}{vector of times of transition to long-term infection for each individual}

\item{R}{vector of times of transition to recovery for each individual}

\item{beta_a}{probability of infection for acute phase}

\item{beta_l}{probability of infection for long-term phase}

\item{Prob_Distr_Params}{probability distribution or mixing matrix for network property of interest}

\item{Network_stats}{network property of interest ("Degree" or "Mixing")}

\item{Prob_Distr}{assumed distribution ("Multinomial_Poisson")}

\item{covPattern}{covariance pattern, use for mixing}
}
\value{
list(CCMnet_Result[\link{1}], CCMnet_Result[\link{2}]), updated contact network
}
\description{
updates the contact network as part of a Metropolis-Hastings algorithm
}
