% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_bayes_inf.R
\name{Update_Il}
\alias{Update_Il}
\title{Update_Il}
\usage{
Update_Il(G, P, Ia, Il, R, beta_a, beta_l, gamma_a, gamma_l)
}
\arguments{
\item{G}{contact network}

\item{P}{transmission network}

\item{Ia}{vector of times of infection for each individual}

\item{Il}{vector of times of transition to long-term infection for each individual}

\item{R}{vector of times of transition to recovery for each individual}

\item{beta_a}{probability of infection for acute phase}

\item{beta_l}{probability of infection for long-term phase}

\item{gamma_a}{length of infection for acute phase}

\item{gamma_l}{length of infection for long-term phase}
}
\value{
new_Il, updated long-term infection times
}
\description{
updates long-term infection times
}
