% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesInf_v4_func_bayes_inf.R
\name{Update_Prob_Distr_Params}
\alias{Update_Prob_Distr_Params}
\title{Update_Prob_Distr_Params}
\usage{
Update_Prob_Distr_Params(
  g,
  Prob_Distr_Params_hyperprior,
  Network_stats,
  Prob_Distr,
  Prob_Distr_Params,
  G_stats,
  MCMC_wgt
)
}
\arguments{
\item{g}{?}

\item{Prob_Distr_Params_hyperprior}{hyperpriors for behavior data}

\item{Network_stats}{?}

\item{Prob_Distr}{?}

\item{Prob_Distr_Params}{?}

\item{G_stats}{?}

\item{MCMC_wgt}{weight from 0 to 1 indicating how much weight to put on the prior versus the MCMC weight}
}
\value{
Prob_Distr_Params parameter values
}
\description{
updates the probability distribution parameters as part of a Metropolis-Hastings algorithm
}
